+++ RELEASE_NOTES.md
@@ -13,6 +13,7 @@
 * GMB-1184 - Four corner logging is not working as expected
 * GMB-1185 - QS: Support for consumer(s) pause/resume
 * GMB-1164: Add an initial pipeline ID info to workResult
+* Property update : Added MAPR ticket file location in g2mailboxservice properties at all environment
 
 5.4.9
 ---------------